refactor: 기기 관리 페이지 리팩토링
